Paar bugfixes mit input values

+++ app/src/main/java/com/example/smartsave/dataClasses/Konto.kt
@@ -5,7 +5,7 @@ import java.time.LocalDate
 
 
 data class Konto(
-    val kontonr: Int,
+    val kontonr: Long,
     val blz: String,
     val bic: String,
     val iban: String,
